+++ Gopkg.toml
@@ -40,6 +40,3 @@
   name = "github.com/streadway/amqp"
   branch = "master"
 
-[[constraint]]
-  name = "github.com/Azure/azure-sdk-for-go"
-  version = "1.2.0"
