+++ agents/dispatcher/main.go
@@ -14,12 +14,12 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/Azure/adx-automation-agent/common"
-	"github.com/Azure/adx-automation-agent/kubeutils"
-	"github.com/Azure/adx-automation-agent/models"
-	"github.com/Azure/adx-automation-agent/monitor"
-	"github.com/Azure/adx-automation-agent/reportutils"
-	"github.com/Azure/adx-automation-agent/schedule"
+	"github.com/Azure/adx-automation-agent/sdk/common"
+	"github.com/Azure/adx-automation-agent/sdk/kubeutils"
+	"github.com/Azure/adx-automation-agent/sdk/models"
+	"github.com/Azure/adx-automation-agent/sdk/monitor"
+	"github.com/Azure/adx-automation-agent/sdk/reportutils"
+	"github.com/Azure/adx-automation-agent/sdk/schedule"
 )
 
 var (
