+++ sdk/models/droidMetadata.go
