+++ sdk/schedule/taskbroker.go
@@ -5,10 +5,10 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/Azure/adx-automation-agent/kubeutils"
+	"github.com/Azure/adx-automation-agent/sdk/kubeutils"
 
-	"github.com/Azure/adx-automation-agent/common"
-	"github.com/Azure/adx-automation-agent/models"
+	"github.com/Azure/adx-automation-agent/sdk/common"
+	"github.com/Azure/adx-automation-agent/sdk/models"
 	"github.com/streadway/amqp"
 )
